+++ lib/main.dart
@@ -9,15 +9,13 @@ import 'package:kmbal_ionicons/kmbal_ionicons.dart';
 import 'package:media_kit/media_kit.dart';
 import 'package:nipaplay/l10n/app_locale_utils.dart';
 import 'package:nipaplay/l10n/app_localizations.dart';
-import 'package:nipaplay/l10n/l10n.dart';
 import 'package:nipaplay/pages/tab_labels.dart';
-import 'package:nipaplay/utils/app_theme.dart';
 import 'package:nipaplay/utils/globals.dart' as globals;
 import 'package:nipaplay/utils/theme_notifier.dart';
 import 'package:nipaplay/utils/system_resource_monitor.dart';
 import 'package:nipaplay/themes/nipaplay/widgets/custom_scaffold.dart';
-import 'package:nipaplay/themes/nipaplay/widgets/menu_button.dart';
-import 'package:nipaplay/themes/nipaplay/widgets/system_resource_display.dart';
+import 'package:nipaplay/themes/nipaplay/widgets/large_screen_mode_actions.dart';
+import 'package:nipaplay/themes/nipaplay/widgets/large_screen_mode_preferences.dart';
 import 'package:window_manager/window_manager.dart';
 import 'package:provider/provider.dart';
 import 'pages/anime_page.dart';
@@ -1090,10 +1088,6 @@ class MainPage extends StatefulWidget {
 
 class MainPageState extends State<MainPage>
     with TickerProviderStateMixin, WindowListener {
-  static const double kWindowCaptionHeight = 28;
-  static const String _largeScreenLayoutPrefKey =
-      'nipaplay_use_large_screen_layout';
-
   bool isMaximized = false;
   TabController? globalTabController;
   bool _showSplash = true;
@@ -1205,7 +1199,7 @@ class MainPageState extends State<MainPage>
   Future<void> _initializeController() async {
     final prefs = await SharedPreferences.getInstance();
     _defaultPageIndex = prefs.getInt('default_page_index') ?? 0;
-    _useLargeScreenLayout = prefs.getBool(_largeScreenLayoutPrefKey) ?? false;
+    _useLargeScreenLayout = await LargeScreenModePreferences.load();
 
     // 强制启用页面滑动动画
     // ... (注释省略)
@@ -1385,8 +1379,7 @@ class MainPageState extends State<MainPage>
       _useLargeScreenLayout = nextValue;
     });
     try {
-      final prefs = await SharedPreferences.getInstance();
-      await prefs.setBool(_largeScreenLayoutPrefKey, nextValue);
+      await LargeScreenModePreferences.save(nextValue);
     } catch (e) {
       debugPrint('[MainPageState] 保存大屏幕模式设置失败: $e');
     }
@@ -1572,420 +1565,27 @@ class MainPageState extends State<MainPage>
             if (!globals.isDesktopOrTablet || !shouldShowAppBar) {
               return const SizedBox.shrink();
             }
-            final String themeActionLabel = isDarkMode
-                ? context.l10n.toggleToLightMode
-                : context.l10n.toggleToDarkMode;
-            final IconData themeActionIcon = isDarkMode
-                ? Icons.nightlight_rounded
-                : Icons.light_mode_rounded;
-
-            final actionButtonsInLargeLayout = Column(
-              mainAxisSize: MainAxisSize.min,
-              children: [
-                NipaplaySidePanel(
-                  isDarkMode: isDarkMode,
-                  child: Column(
-                    mainAxisSize: MainAxisSize.min,
-                    children: [
-                      NipaplaySidePanelItem(
-                        isSelected: false,
-                        activeColor: const Color(0xFFFF2E55),
-                        inactiveColor:
-                            isDarkMode ? Colors.white60 : Colors.black54,
-                        onTap: _toggleLargeScreenLayout,
-                        child: Row(
-                          children: const [
-                            Icon(Icons.view_day_rounded, size: 20),
-                            SizedBox(width: 8),
-                            Expanded(
-                              child: Text(
-                                '退出大屏幕模式',
-                                maxLines: 1,
-                                overflow: TextOverflow.ellipsis,
-                                style: TextStyle(
-                                  fontSize: 16,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      NipaplaySidePanelItem(
-                        isSelected: false,
-                        activeColor: const Color(0xFFFF2E55),
-                        inactiveColor:
-                            isDarkMode ? Colors.white60 : Colors.black54,
-                        onTap: () {
-                          context.read<ThemeNotifier>().themeMode =
-                              _nextThemeMode();
-                        },
-                        child: Row(
-                          children: [
-                            Icon(themeActionIcon, size: 20),
-                            const SizedBox(width: 8),
-                            Expanded(
-                              child: Text(
-                                themeActionLabel,
-                                maxLines: 1,
-                                overflow: TextOverflow.ellipsis,
-                                style: const TextStyle(
-                                  fontSize: 16,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                      NipaplaySidePanelItem(
-                        isSelected: false,
-                        activeColor: const Color(0xFFFF2E55),
-                        inactiveColor:
-                            isDarkMode ? Colors.white60 : Colors.black54,
-                        onTap: () => SettingsPage.showWindow(context),
-                        child: Row(
-                          children: [
-                            const Icon(Icons.settings_rounded, size: 20),
-                            const SizedBox(width: 8),
-                            Expanded(
-                              child: Text(
-                                context.l10n.settingsLabel,
-                                maxLines: 1,
-                                overflow: TextOverflow.ellipsis,
-                                style: const TextStyle(
-                                  fontSize: 16,
-                                  fontWeight: FontWeight.bold,
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            );
-            final actionButtonsInNormalLayout = Row(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                SizedBox(
-                  height: kWindowCaptionHeight,
-                  child: Center(
-                    child: Image.asset(
-                      'assets/logo2.png',
-                      height: 24,
-                      fit: BoxFit.contain,
-                      color: isDarkMode ? Colors.white : Colors.black,
-                      colorBlendMode: BlendMode.srcIn,
-                    ),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                SizedBox(
-                  height: kWindowCaptionHeight,
-                  child: Center(
-                    child: _LargeScreenLayoutToggleButton(
-                      isActive: isLargeScreenLayoutActive,
-                      onPressed: _toggleLargeScreenLayout,
-                    ),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                SizedBox(
-                  height: kWindowCaptionHeight,
-                  child: Center(
-                    child: _ThemeToggleButton(
-                      onToggleFromOrigin: _handleThemeToggleFromButton,
-                    ),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                SizedBox(
-                  height: kWindowCaptionHeight,
-                  child: Center(
-                    child: _SettingsEntryButton(
-                      onPressed: () => SettingsPage.showWindow(context),
-                    ),
-                  ),
-                ),
-              ],
-            );
-            return Positioned(
-              top: isLargeScreenLayoutActive ? null : topPadding,
-              left: isLargeScreenLayoutActive ? 0 : null,
-              right: isLargeScreenLayoutActive ? null : rightPadding,
-              bottom: isLargeScreenLayoutActive
-                  ? 0
-                  : null,
-              child: isLargeScreenLayoutActive
-                  ? actionButtonsInLargeLayout
-                  : Row(
-                      mainAxisSize: MainAxisSize.min,
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        const SystemResourceDisplay(),
-                        const SizedBox(width: 8),
-                        actionButtonsInNormalLayout,
-                        const SizedBox(width: 8),
-                        if (!kIsWeb && (Platform.isWindows || Platform.isLinux))
-                          SizedBox(
-                            height: kWindowCaptionHeight,
-                            child: Center(
-                              child: WindowControlButtons(
-                                isMaximized: isMaximized,
-                                onMinimize: _minimizeWindow,
-                                onMaximizeRestore: _toggleWindowSize,
-                                onClose: _closeWindow,
-                              ),
-                            ),
-                          ),
-                      ],
-                    ),
+            return NipaplayLargeScreenModeActionsOverlay(
+              isDarkMode: isDarkMode,
+              isLargeScreenLayoutActive: isLargeScreenLayoutActive,
+              topPadding: topPadding,
+              rightPadding: rightPadding,
+              showWindowsButtons:
+                  !kIsWeb && (Platform.isWindows || Platform.isLinux),
+              isMaximized: isMaximized,
+              onToggleLargeScreen: _toggleLargeScreenLayout,
+              onToggleThemeFromOrigin: _handleThemeToggleFromButton,
+              onOpenSettings: () => SettingsPage.showWindow(context),
+              onMinimize: _minimizeWindow,
+              onMaximizeRestore: _toggleWindowSize,
+              onClose: _closeWindow,
             );
           },
         ),
-        if (isLargeScreenLayoutActive &&
-            !kIsWeb &&
-            (Platform.isWindows || Platform.isLinux))
-          Positioned(
-            top: topPadding,
-            right: rightPadding,
-            child: SizedBox(
-              height: kWindowCaptionHeight,
-              child: Center(
-                child: WindowControlButtons(
-                  isMaximized: isMaximized,
-                  onMinimize: _minimizeWindow,
-                  onMaximizeRestore: _toggleWindowSize,
-                  onClose: _closeWindow,
-                ),
-              ),
-            ),
-          ),
       ],
     );
 
     return content;
-  }
-}
-
-class _SettingsEntryButton extends StatefulWidget {
-  final VoidCallback onPressed;
-
-  const _SettingsEntryButton({required this.onPressed});
-
-  @override
-  State<_SettingsEntryButton> createState() => _SettingsEntryButtonState();
-}
-
-class _LargeScreenLayoutToggleButton extends StatefulWidget {
-  const _LargeScreenLayoutToggleButton({
-    required this.isActive,
-    required this.onPressed,
-  });
-
-  final bool isActive;
-  final VoidCallback onPressed;
-
-  @override
-  State<_LargeScreenLayoutToggleButton> createState() =>
-      _LargeScreenLayoutToggleButtonState();
-}
-
-class _ThemeToggleButton extends StatefulWidget {
-  final Future<void> Function(Offset globalOrigin)? onToggleFromOrigin;
-
-  const _ThemeToggleButton({this.onToggleFromOrigin});
-
-  @override
-  State<_ThemeToggleButton> createState() => _ThemeToggleButtonState();
-}
-
-class _ThemeToggleButtonState extends State<_ThemeToggleButton> {
-  bool _isHovered = false;
-  bool _isPressed = false;
-
-  void _setHovered(bool value) {
-    if (_isHovered == value) {
-      return;
-    }
-    setState(() {
-      _isHovered = value;
-    });
-  }
-
-  void _toggleTheme() {
-    final onToggleFromOrigin = widget.onToggleFromOrigin;
-    if (onToggleFromOrigin != null) {
-      final renderObject = context.findRenderObject();
-      if (renderObject is RenderBox && renderObject.hasSize) {
-        final origin =
-            renderObject.localToGlobal(renderObject.size.center(Offset.zero));
-        unawaited(onToggleFromOrigin(origin));
-        return;
-      }
-    }
-
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
-    context.read<ThemeNotifier>().themeMode =
-        isDarkMode ? ThemeMode.light : ThemeMode.dark;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
-    final double scale = _isPressed ? 0.92 : (_isHovered ? 1.1 : 1.0);
-    final Color iconColor = _isHovered
-        ? const Color(0xFFFF2E55)
-        : (isDarkMode ? Colors.white : Colors.black87);
-    final icon =
-        isDarkMode ? Icons.nightlight_rounded : Icons.light_mode_rounded;
-    final tooltip = isDarkMode
-        ? context.l10n.toggleToLightMode
-        : context.l10n.toggleToDarkMode;
-
-    return Tooltip(
-      message: tooltip,
-      child: MouseRegion(
-        onEnter: (_) => _setHovered(true),
-        onExit: (_) => _setHovered(false),
-        child: GestureDetector(
-          onTapDown: (_) => setState(() => _isPressed = true),
-          onTapUp: (_) => setState(() => _isPressed = false),
-          onTapCancel: () => setState(() => _isPressed = false),
-          onTap: _toggleTheme,
-          child: AnimatedScale(
-            scale: scale,
-            duration: const Duration(milliseconds: 120),
-            child: AnimatedSwitcher(
-              duration: const Duration(milliseconds: 320),
-              switchInCurve: Curves.easeOutCubic,
-              switchOutCurve: Curves.easeInCubic,
-              transitionBuilder: (child, animation) {
-                return FadeTransition(
-                  opacity: animation,
-                  child: ScaleTransition(
-                    scale:
-                        Tween<double>(begin: 0.85, end: 1.0).animate(animation),
-                    child: RotationTransition(
-                      turns: Tween<double>(begin: 0.9, end: 1.0)
-                          .animate(animation),
-                      child: child,
-                    ),
-                  ),
-                );
-              },
-              child: Icon(
-                icon,
-                key: ValueKey<bool>(isDarkMode),
-                size: 22,
-                color: iconColor,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _LargeScreenLayoutToggleButtonState
-    extends State<_LargeScreenLayoutToggleButton> {
-  bool _isHovered = false;
-  bool _isPressed = false;
-
-  void _setHovered(bool value) {
-    if (_isHovered == value) {
-      return;
-    }
-    setState(() {
-      _isHovered = value;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
-    final bool isActive = widget.isActive;
-    final double scale = _isPressed ? 0.92 : (_isHovered ? 1.1 : 1.0);
-    final Color iconColor = isActive
-        ? const Color(0xFFFF2E55)
-        : (_isHovered
-            ? const Color(0xFFFF2E55)
-            : (isDarkMode ? Colors.white : Colors.black87));
-    final icon = isActive ? Icons.view_day_rounded : Icons.view_sidebar_rounded;
-
-    return Tooltip(
-      message: isActive ? '退出大屏幕模式' : '大屏幕模式',
-      child: MouseRegion(
-        onEnter: (_) => _setHovered(true),
-        onExit: (_) => _setHovered(false),
-        child: GestureDetector(
-          onTapDown: (_) => setState(() => _isPressed = true),
-          onTapUp: (_) => setState(() => _isPressed = false),
-          onTapCancel: () => setState(() => _isPressed = false),
-          onTap: widget.onPressed,
-          child: AnimatedScale(
-            scale: scale,
-            duration: const Duration(milliseconds: 120),
-            child: Icon(
-              icon,
-              size: 22,
-              color: iconColor,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class _SettingsEntryButtonState extends State<_SettingsEntryButton> {
-  bool _isHovered = false;
-  bool _isPressed = false;
-
-  void _setHovered(bool value) {
-    if (_isHovered == value) {
-      return;
-    }
-    setState(() {
-      _isHovered = value;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final isDarkMode = Theme.of(context).brightness == Brightness.dark;
-    final double scale = _isPressed ? 0.92 : (_isHovered ? 1.1 : 1.0);
-    final Color iconColor = _isHovered
-        ? const Color(0xFFFF2E55)
-        : (isDarkMode ? Colors.white : Colors.black87);
-
-    return Tooltip(
-      message: context.l10n.settingsLabel,
-      child: MouseRegion(
-        onEnter: (_) => _setHovered(true),
-        onExit: (_) => _setHovered(false),
-        child: GestureDetector(
-          onTapDown: (_) => setState(() => _isPressed = true),
-          onTapUp: (_) => setState(() => _isPressed = false),
-          onTapCancel: () => setState(() => _isPressed = false),
-          onTap: widget.onPressed,
-          child: AnimatedScale(
-            scale: scale,
-            duration: const Duration(milliseconds: 120),
-            child: Icon(
-              Icons.settings_rounded,
-              size: 22,
-              color: iconColor,
-            ),
-          ),
-        ),
-      ),
-    );
   }
 }
